Rewrite of fnpf printer like for ecb8ef19
