+++ src/REEF3D.project
@@ -424,12 +424,7 @@
             <File Name="fnpf_vtp_fsf.h"/>
           </VirtualDirectory>
           <File Name="fnpf_printer.h"/>
-          <VirtualDirectory Name="VTU">
-            <File Name="fnpf_vtu3D.h"/>
-          </VirtualDirectory>
-          <VirtualDirectory Name="VTS">
-            <File Name="fnpf_vts3D.h"/>
-          </VirtualDirectory>
+          <File Name="printer_fnpf.h"/>
           <VirtualDirectory Name="Probes">
             <File Name="fnpf_vel_probe.h"/>
             <File Name="fnpf_vel_probe_theory.h"/>
@@ -1609,21 +1604,13 @@
             <File Name="fnpf_breaking_log_filename.cpp"/>
             <File Name="fnpf_breaking_log.cpp"/>
           </VirtualDirectory>
-          <File Name="fnpf_vtu3D.cpp"/>
-          <File Name="fnpf_vts3D.cpp"/>
+          <File Name="printer_fnpf.cpp"/>
+          <File Name="printer_fnpf_parallel.cpp"/>
           <VirtualDirectory Name="WSF">
             <File Name="fnpf_print_wsfline_y.cpp"/>
             <File Name="fnpf_print_wsfline.cpp"/>
             <File Name="fnpf_print_wsf_theory.cpp"/>
             <File Name="fnpf_print_wsf.cpp"/>
-          </VirtualDirectory>
-          <VirtualDirectory Name="VTU">
-            <File Name="fnpf_vtu3D_pvtu.cpp"/>
-            <File Name="fnpf_vtu3D_header.cpp"/>
-          </VirtualDirectory>
-          <VirtualDirectory Name="VTS">
-            <File Name="fnpf_vts3D_pvts.cpp"/>
-            <File Name="fnpf_vts3D_header.cpp"/>
           </VirtualDirectory>
           <VirtualDirectory Name="VTP FSF">
             <File Name="fnpf_vtp_fsf_pvtp.cpp"/>
