+++ src/printer_fnpf.h
@@ -22,6 +22,7 @@ Author: Hans Bihs
 
 #include"fnpf_printer.h"
 #include"increment.h"
+#include "vtks.h"
 
 class fdm_fnpf;
 class force_ale;
@@ -45,21 +46,18 @@ class ioflow;
 
 using namespace std;
 
-class fnpf_vtu3D : public fnpf_printer, public increment
+class printer_fnpf : public fnpf_printer, public increment
 {
 
 public:
-	fnpf_vtu3D(lexer*,fdm_fnpf*,ghostcell*);
-	virtual ~fnpf_vtu3D();
+	printer_fnpf(lexer*,fdm_fnpf*,ghostcell*);
+	virtual ~printer_fnpf();
 	virtual void start(lexer*,fdm_fnpf*,ghostcell*,ioflow*);
     virtual void print_stop(lexer*,fdm_fnpf*,ghostcell*);
     virtual void print_vtk(lexer*,fdm_fnpf*,ghostcell*);
     
 private:
-    void pvtk(lexer*,ghostcell*);
-    void name_iter(lexer*,ghostcell*);
-    void name_time(lexer*,ghostcell*);
-    void piecename(lexer*,ghostcell*, int);
+    void parallelData(lexer*,ghostcell*);
 
     char name[200],pname[200],epsvar[200];
     int n,iin,offset[200];
@@ -87,6 +85,8 @@ private:
     fnpf_print_Hs *phs;
     fnpf_vel_probe *pvel;
     fnpf_vel_probe_theory *pveltheo;
+
+    vtk3D *outputFormat;
 };
 
 #endif
