+++ src/vtk3D.h
@@ -27,6 +27,7 @@ Author: Hans Bihs
 
 class lexer;
 class fdm;
+class fdm_fnpf;
 class ghostcell;
 
 class vtk3D
@@ -35,6 +36,7 @@ class vtk3D
         virtual void folder(char*){};
         virtual void offset(lexer*,int*,int&){};
         virtual void structureWrite(lexer*, fdm*, std::ofstream&){};
+        virtual void structureWrite(lexer*, fdm_fnpf*, std::ofstream&){};
         virtual void extent(lexer*,ghostcell*){};
         virtual void fileName(char*, char*, int&, int&){};
         virtual void parallelFileName(char*, char*, int&){};
