+++ src/vtu3D.cpp
@@ -23,6 +23,7 @@ Author: Hans Bihs
 #include "vtu3D.h"
 #include "lexer.h"
 #include "fdm.h"
+#include "fdm_fnpf.h"
 #include <sys/stat.h>
 
 vtu3D::vtu3D()
@@ -148,7 +149,6 @@ void vtu3D::structureWrite(lexer *p, fdm *a, std::ofstream &result)
 					+ p->B183_1*sin((2.0*PI*p->B183_2)*p->simtime + p->B183_3));
 		result.write((char*)&ffn, sizeof (float));
 	}
-
 	//  Connectivity
     iin=4*(p->tpcellnum)*8;
     result.write((char*)&iin, sizeof (int));
@@ -180,7 +180,88 @@ void vtu3D::structureWrite(lexer *p, fdm *a, std::ofstream &result)
 		result.write((char*)&iin, sizeof (int));
 	}
 
-	// //  Offset of Connectivity
+	structureWriteEnd(p,result);
+}
+
+void vtu3D::structureWrite(lexer *p, fdm_fnpf *c, std::ofstream &result)
+{
+	float ffn;
+	int iin;
+	double phase=0.0;
+	double zcoor;
+
+	//  XYZ
+	double theta_y = p->B192_1*(PI/180.0);
+	double omega_y = 2.0*PI*p->B192_2;
+	double waterlevel;
+
+	if(p->B192==1 && p->simtime>=p->B194_s && p->simtime<=p->B194_e)
+	phase = omega_y*p->simtime;
+
+	iin=4*(p->pointnum)*3;
+	result.write((char*)&iin, sizeof (int));
+	TPLOOP
+	{
+		waterlevel = p->sl_ipol4eta(p->wet,c->eta,c->bed)+p->wd - p->sl_ipol4(c->bed);
+
+		zcoor = p->ZN[KP1]*waterlevel + p->sl_ipol4(c->bed);
+
+
+		if(p->wet[IJ]==0)
+		zcoor=c->bed(i,j);
+
+		if(i+p->origin_i==-1 && j+p->origin_j==-1 && p->wet[(0-p->imin)*p->jmax + (0-p->jmin)]==1)
+		zcoor = p->ZN[KP1]*c->WL(i,j) + c->bed(i,j);
+
+		ffn=float((p->XN[IP1]-p->B192_3)*cos(theta_y*sin(phase)) - (zcoor-p->B192_4)*sin(theta_y*sin(phase)) + p->B192_3);
+		result.write((char*)&ffn, sizeof (float));
+
+		ffn=float(p->YN[JP1]);
+		result.write((char*)&ffn, sizeof (float));
+
+		ffn=float((p->XN[IP1]-p->B192_3)*sin(theta_y*sin(phase)) + (zcoor-p->B192_4)*cos(theta_y*sin(phase)) + p->B192_4);
+		result.write((char*)&ffn, sizeof (float));
+	}
+
+    //  Connectivity
+	iin=4*(p->tpcellnum)*8;
+	result.write((char*)&iin, sizeof (int));
+	BASELOOP
+	if(p->flag5[IJK]!=-20 && p->flag5[IJK]!=-30)
+	{
+		iin=int(c->nodeval(i-1,j-1,k-1)-1);
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i,j-1,k-1))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin= int(c->nodeval(i,j,k-1))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i-1,j,k-1))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i-1,j-1,k))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i,j-1,k))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i,j,k))-1;
+		result.write((char*)&iin, sizeof (int));
+
+		iin=int(c->nodeval(i-1,j,k))-1;
+		result.write((char*)&iin, sizeof (int));
+	}
+
+	structureWriteEnd(p,result);
+}
+
+void vtu3D::structureWriteEnd(lexer *p, std::ofstream &result)
+{
+	int iin;
+
+	// Offset of Connectivity
     iin=4*(p->tpcellnum);
     result.write((char*)&iin, sizeof (int));
 	for(n=0;n<p->tpcellnum;++n)
