+++ src/vtu3D.h
@@ -26,9 +26,6 @@ Author: Hans Bihs
 #include "vtk3D.h"
 #include "increment.h"
 
-class lexer;
-class fdm;
-
 class vtu3D : public vtk3D , increment
 {
     public:
@@ -38,6 +35,7 @@ class vtu3D : public vtk3D , increment
         void folder(char*);
         void offset(lexer*, int*, int&);
         void structureWrite(lexer*, fdm*, std::ofstream&);
+        void structureWrite(lexer*, fdm_fnpf*, std::ofstream&);
 
         void beginning(lexer*, std::ofstream&);
         void beginningParallel(lexer*, std::ofstream&);
@@ -45,6 +43,8 @@ class vtu3D : public vtk3D , increment
         void endingParallel(std::ofstream&, char*, int&, int&);
         void fileName(char *name, char *A10, int &num, int &rank){sprintf(name,"./REEF3D_%s_VTU/REEF3D-%s-%08i-%06i.vtu",A10,A10,num,rank);};
         void parallelFileName(char *name, char *A10, int &num){sprintf(name,"./REEF3D_%s_VTU/REEF3D-%s-%08i.pvtu",A10,A10,num);};
+    private:
+        void structureWriteEnd(lexer*, std::ofstream&);
 };
 
 #endif
