+++ src/waves_header.h
@@ -31,9 +31,5 @@ Author: Hans Bihs
 #include"fnpf_RK3.h"
 #include"fnpf_RK4.h"
 #include"fnpf_printer.h"
-#include"fnpf_vtu3D.h"
-#include"fnpf_vts3D.h"
+#include"printer_fnpf.h"
 #include"fnpf_timestep.h"
-
-
-
